Split CLI out to files per feature.
